requirements.txt
psycopg2-binary>=2.9
sqlite-vec>=0.1
weaviate-client>=4.0
pinecone>=8.0
numpy>=1.26
tqdm>=4.66
jupyter>=1.0
nbformat>=5.9
pandas>=2.1
matplotlib>=3.8
python-dotenv>=1.0
pytest>=8.0
